+++ core/src/commonMain/kotlin/com/powersync/db/PowerSyncDatabaseImpl.kt
@@ -1,7 +1,9 @@
 package com.powersync.db
 
+import androidx.sqlite.SQLiteConnection
 import co.touchlab.kermit.Logger
 import com.powersync.DatabaseDriverFactory
+import com.powersync.ExperimentalPowerSyncAPI
 import com.powersync.PowerSyncDatabase
 import com.powersync.PowerSyncException
 import com.powersync.bucket.BucketPriority
@@ -90,6 +92,7 @@ internal class PowerSyncDatabaseImpl(
             dbFilename = dbFilename,
             dbDirectory = dbDirectory,
             writeLockMutex = resource.group.writeLockMutex,
+            logger = logger,
         )
 
     internal val bucketStorage: BucketStorage = BucketStorageImpl(internalDb, logger)
@@ -313,6 +316,12 @@ internal class PowerSyncDatabaseImpl(
         // The initialization sets powerSyncVersion.
         waitReady()
         return powerSyncVersion
+    }
+
+    @ExperimentalPowerSyncAPI
+    override suspend fun leaseConnection(readOnly: Boolean): SQLiteConnection {
+        waitReady()
+        return internalDb.leaseConnection(readOnly)
     }
 
     override suspend fun <RowType : Any> get(